# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoFinder HTTP API",
    "version": "1.0",
    "description": "Stable same-origin authenticated API. Mutations also require X-CSRF-TOKEN."
  },
  "servers": [{"url": "/sofinder"}],
  "security": [{"sessionCookie": []}],
  "paths": {
...
    "/api/folders": {"post": {"operationId": "createFolder", "parameters": [{"$ref": "#/components/parameters/csrf"}], "requestBody": {"$ref": "#/components/requestBodies/EntryName"}, "responses": {"201": {"$ref": "#/components/responses/EntryResult"}}}},
...
    "/api/trash/{id}/restore": {"post": {"operationId": "restoreTrash", "parameters": [{"$ref": "#/components/parameters/id"}, {"$ref": "#/components/parameters/csrf"}], "requestBody": {"$ref": "#/components/requestBodies/Trash"}, "responses": {"200": {"$ref": "#/components/responses/EntryResult"}}}},
...
    "/api/metadata": {
      "get": {"operationId": "getMetadata", "parameters": [{"$ref": "#/components/parameters/resource"}], "responses": {"200": {"$ref": "#/components/responses/Metadata"}}},
      "patch": {"operationId": "updateMetadata", "parameters": [{"$ref": "#/components/parameters/csrf"}], "requestBody": {"$ref": "#/components/requestBodies/Metadata"}, "responses": {"200": {"$ref": "#/components/responses/Metadata"}}}
    },
...
  },
  "components": {
    "securitySchemes": {"sessionCookie": {"type": "apiKey", "in": "cookie", "name": "PHPSESSID"}},
    "parameters": {
      "resource": {"name": "resource", "in": "query", "schema": {"type": "string", "default": "Files"}},
...
      "id": {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}},
      "csrf": {"name": "X-CSRF-TOKEN", "in": "header", "required": true, "schema": {"type": "string"}}
    },
    "requestBodies": {
...
      "EntryName": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EntryNameRequest"}}}},
...
      "Trash": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TrashRequest"}}}},
...
      "Metadata": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MetadataRequest"}}}},
...
    },
    "schemas": {
...
      "AssetReference": {"$ref": "./schema/asset-reference.schema.json"},
...
      "Entry": {"type": "object", "additionalProperties": false, "required": ["path", "name", "directory", "size", "modifiedAt", "mimeType", "url", "capabilities"], "properties": {"path": {"type": "string"}, "name": {"type": "string"}, "directory": {"type": "boolean"}, "size": {"type": "integer", "minimum": 0}, "modifiedAt": {"type": "integer", "minimum": 0}, "mimeType": {"type": ["string", "null"]}, "url": {"type": ["string", "null"]}, "capabilities": {"type": "object", "additionalProperties": {"type": "boolean"}}}},
      "EntryResult": {"type": "object", "additionalProperties": false, "required": ["entry"], "properties": {"entry": {"$ref": "#/components/schemas/Entry"}, "asset": {"oneOf": [{"$ref": "#/components/schemas/AssetReference"}, {"type": "null"}]}}},
...
      "MetadataData": {"type": "object", "additionalProperties": false, "required": ["favorites", "quickAccess", "quickAccessEntries", "tags", "recent"], "properties": {"favorites": {"type": "array", "items": {"type": "string"}}, "quickAccess": {"type": "array", "maxItems": 12, "items": {"type": "string"}}, "quickAccessEntries": {"type": "array", "maxItems": 12, "items": {"type": "object", "additionalProperties": false, "required": ["path", "name", "directory", "mimeType", "exists"], "properties": {"path": {"type": "string"}, "name": {"type": "string"}, "directory": {"type": ["boolean", "null"]}, "mimeType": {"type": ["string", "null"]}, "exists": {"type": "boolean"}}}}, "tags": {"type": "object", "additionalProperties": {"type": "array", "items": {"type": "string"}}}, "recent": {"type": "array", "items": {"type": "object", "additionalProperties": false, "required": ["path", "touchedAt"], "properties": {"path": {"type": "string"}, "touchedAt": {"type": "integer"}}}}}},
...
      "EntryNameRequest": {"type": "object", "additionalProperties": false, "required": ["resource", "path", "name"], "properties": {"resource": {"type": "string", "minLength": 1}, "path": {"type": "string"}, "name": {"type": "string", "minLength": 1}}},
...
      "TrashRequest": {"type": "object", "additionalProperties": false, "required": ["resource"], "properties": {"resource": {"type": "string"}, "conflict": {"enum": ["cancel", "overwrite", "rename"]}}},
...
      "MetadataRequest": {"type": "object", "additionalProperties": false, "required": ["resource", "path", "action"], "properties": {"resource": {"type": "string"}, "path": {"type": "string"}, "action": {"enum": ["favorite", "quick_access", "tags", "touch", "forget"]}, "favorite": {"type": "boolean"}, "pinned": {"type": "boolean"}, "tags": {"type": "array", "uniqueItems": true, "items": {"type": "string"}}}},
...
      "EntryResultEnvelope": {"type": "object", "additionalProperties": false, "required": ["success", "data"], "properties": {"success": {"const": true}, "data": {"$ref": "#/components/schemas/EntryResult"}}},
...
      "MetadataEnvelope": {"type": "object", "additionalProperties": false, "required": ["success", "data"], "properties": {"success": {"const": true}, "data": {"$ref": "#/components/schemas/MetadataData"}}},
...
    },
    "responses": {
...
      ,"EntryResult": {"description": "Typed entry and optional editor asset", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EntryResultEnvelope"}}}}
...
      ,"Metadata": {"description": "User-scoped file metadata", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MetadataEnvelope"}}}}
    }
  }
}